New functions in Track subclasses report type info...
... A virtual function GetTypeInfo for run-time polymorphism, and a static
ClassTypeInfo in each sub-class, for compile-time polymorphism in templates.

This is the first step to open-endedness in the definition of track subclasses,
eliminating the need for a centralized list of all subtypes (not even as
incomplete types).

(modified from Audacity commit 5a22a5f8b174d75403ee38a5a6683264a0e63614)
Signed-off-by: Avery King <avery98@pm.me>