Rename ScrubState to ScrubQueue and make it a singleton...
... and move some static functions into a struct now called ScrubState

This makes a cycle between ScrubState and AudioIO but it will be broken soon.

Notice the deleted `mScrubState.reset();` in AudioIO::StopStream.  Removing it
is mostly harmless, but the effect will be restored anyway in the next commit,
as it is moved into ScrubState::Finalize().

(cherry picked from commit d5418c88aabd44813294685f5ccc7e756a63730f)
Signed-off-by: Avery King <avery98@pm.me>