FreqWindow: Implement horizontal zoom/scroll
This uses an exponential curve when zooming horizontally. This feels
more natural than the "reciprocal" curve used for vertical zoom. I tried
using that curve for horizontal zoom as well, but it was flawed because
it would zoom slowly at first but accelerate as you approached maximum
zoom level.

One minor issue is that dragging the zoom slider now results in zooming
in slightly to the left of the center of the plot. This is because every
change in zoom level recomputes the horizontal scrollbar's size and
position, rounding the position down to the nearest integer. The
rounding errors accumulate over many updates.

This needs to be solved by storing the intended center/boundaries of the
viewport (either its frequency or position) independently of the
scrollbar.

Signed-off-by: nyanpasu64 <nyanpasu64@tuta.io>