Added new 'IsSaucedacity' preference
This preference checks if the created preferences were made by
Saucedacity. If they weren't, the user is alerted, and the situation
corrected. This means very little (virtually nothing) right now, but
this is intended for future proofing

Note: Saucedacity preferences created prior to 1.1 will be affected. As
such, you don't need to worry about it.