Removed "Check for updates..." menu option
Help > "Check for updates..." has been removed. This comes as a decision
to prevent confusion from those coming from Audacity. Additionally, this
option merely opened "https://github.com/saucedacity/saucedacity" in a
web browser, so it wasn't much in the first place.

Although the menu option has been removed, the rest of the code behind
this menu still remains. Because of this, that code will have to stay
for now, although it will be removed eventually.