Add dialog warning if pre-release, name changes
The splash dialog now shows a pre-release warning dialog first if the
version of Saucedacity is either an alpha or beta release. If the user
chooses the option "Don't show this again at startup" in the splash
dialog, this dialog will no longer display. Additionally, the text
warning about an alpha/beta in the splash dialog has been removed as the
other dialog does that for now.

On the other hand, there's been a name change with the splash dialog
title itself. This has been changed from Audacity to Saucedacity.