Initial commit for GenericUI and GenericObject
This begins the transition to a toolkit neutral state. All GUI code is
planed to be implemented (almost) entirely here. All of this will be
under a toolkit wrapper named 'GenericUI'. Additionally, any existing
code with BasicUI::* and friends will be re-used (maybe moved to here,
we'll have to see).

GenericUI isn't actually used in Saucedacity yet. We'll still
implementing some basic things and defining some core concepts of this
wrapper.

Miscellaneous:

* Removed Sentry reference in wxWidgetsBasicUI.cpp. Nothing harmful as
  it was neutered, but if you are alergic to dead code, then this might
  have been an issue. (It kind of is to me, but at least it wasn't a
  huge part of Saucedacity).