Changed "Light" theme to "Audacity"
The light theme is now known as the "Audacity" theme since this was
Audacity's default theme. A new theme will be introduced, replacing the
light theme as the default theme.

This theme will still be kept, but it will not be the default anymore.
For those wanting Audacity's default theme in Saucedacity, however, this
will still remain an option.

Miscellaneous notes:

* ThemePrefs is now shown. This makes it easier to create a built-in
  theme for Saucedacity (and Audacity, for that matter).

* The theming system has some planned changes:
  * Some images will be loaded at runtime rather than be built-in
    instead. This allows for greater customization.

  * The ImageCache format will soon be replaced with a new format, of
    which is described by a config file and has images that are to be
    loaded at runtime by Saucedacity. If loading an external theme
    fails, however, a built-in theme will be used instead.