issue #1838: Screen readers should read labels when tabbing
Now that when a label track is the focus tab and shift+tab only "select" a label, rather than opening it for editing in a custom text box which is not accessible for screen readers, it's useful for the name of the label to be read.

Straightforward fix.
Also added i18n-hint in TransportMenus.cpp where there is the same message for screen readers.