Adds an ability to change project's user_version based on features used
This commit introduces a new mechanism to allow projects components to specify the Audacity version they require to be loaded.

It is possible to set the version based on the features that are currently in use.

Smart Clips use this new approach to require at least Audacity 3.1 if there are trimmed clips in the project.

===============================================================================

Cherry-picked from Audacity commit afb4f24. Modified for Tenacity.

* We can also use these features in case we introduce our own
  Tenacity-specific features for the AUP3 file format in the future.

Signed-off-by: Avery King <avery98@pm.me>