Change shortcut for renaming label to hardcoded Ctrl+F2
Fixes the issue #1710 for label tracks. (And issue #1780, see below).

Problem: For label tracks Enter (hardcoded) for renaming the selected label, is the same shortcut used by default by the command for toggling the selectedness of the track.

Fix: Use the shortcut Ctrl+F2 for renaming the selected label.
WARNING: This has been implemented by hardcoding this shortcut, rather than providing a context sensitive command for renaming both clips and labels. This was done due to a shortage of time. This hardcoded shortcut should be removed after the release of 3.1.0

In changing the shortcut from Enter to Ctrl+F2, in LabelTrackView::DoCaptureKey(), in the line:
if (IsValidIndex(mTextEditIndex, project) || IsValidIndex(mNavigationIndex, project)) {

the case IsValidIndex(mNavigationIndex, project) was removed. This had been added in commit a0ad72d, presumably to capture the Enter key. This was an incorrect change and caused issue #1780:A label can be created by typing regardless of option setting.
So this commit also fixes that issue.