Drop obsolete code
Some obsolete workarounds for
https://bugzilla.audacityteam.org/show_bug.cgi?id=458 were dropped in
PrefsDialog.cpp and EditMenus.cpp. This was because Saucedacity requires
wxWidgets 3.1 (Audacity's fork), thus making these workarounds obsolete.

Additionally, some CB code for Linux was removed. It was ultimately too
out of date to keep anymore, so this was ultimately dropped as well.
(Actually, I've been wanting to do this for some time now).