Writers of XML sub-objects must execute after attributes...
... when they are attached via the registry.  Do not tolerate cross-platform
permutations of the call sequence that violate this requirement!  Still
tolerating permutations that do not violate it.

This system still depends on the registered functions honoring the contract.  It
doesn't check it.

So among existing registered writes, the one for Tags (only) needs to be
guaranteed last.

(cherry picked from Audacity commit 4b8b269; modified for Tenacity).
Signed-off-by: Avery King <avery98@pm.me>