Added notes to some options
This namely adds a notice for the 'DA' option, which enables
DarkAudacity customizations. We might integrate some DarkAudacity
customizations, then drop this option, but I cannot be sure if 1) we'll
integrate any options in the first place and 2) what customizations, if
applicable, would we add.

Additionally, there was a 'NOTEBOOK' option, which was likely added
sometime in 2010 (as far as I can tell). I have no idea what this option
does, but I might rewrite some parts of this code to see what it does.