Reverted saucedacity_enable_legacy, rename all options
This option has been reverted as it was of little significance. It was
just another name for audacity_use_ladspa.

On the other hand, all options are now prefixed with 'saucedacity_'
instead of 'audacity_'. This is to better separate Saucedacity build
options from Audacity build options. (This is also in contrast to the
prior approach of introducing new options with the prefix 'audacity_').