FFmpeg support reorganization
This FFmpeg rework is intended to help better organize the codebase. In
addition, some error messages now display the error message returned by
the corresponding libav function.

This is only part 1 of a series of a larger project: to clean up support
for FFmpeg in Saucedacity. The intention is to fix deprecation warnings
and to fix bugs, of which one was found while reworking FFmpeg support.

The changes contained in this commit were originally intended to fix
some deprecation warnings in AVPacketEx. However, fixing these warnings
appears to have uncovered a bug in which av_encode_audio2 returns
AVERROR(EINVAL). This was tested with the OGG format with the libvorbis
and flac codec options. More testing might be needed just to ensure this
issue is persistent across format and codec options.