Issue2082: move real-time effect calculation to producer side...
... of the playback RingBuffer.

It may be making calls to unknown foreign code in plug-ins.  Can we trust
unknown code really to follow the constraints of a low-latency thread (such
as not allocating memory or locking a mutex or similar)?  Rather let's do it
in the thread that must maintain high throughput, but can tolerate more variance
in processing times of batches of data.

Note that the "drop quickly" logic on the consumer side is only about the
"micro-fades" that last milliseconds, and smooth the transition to silence when
stopping.  Whether those few microseconds are now realtime-transformed, where
they weren't before -- matters little.

Note that our own code does some mutex locking in
RealtimeEffectManager::ProcessScope, constructing and destroying it, and in
Process(); not held for the duraction of the scope.  Some way to avoid this
locking needs to be found later.

This will cause response of playback to changes of the realtime effect settings
to become much more laggy.  This will be fixed later.

(cherry picked from commit 713ced281e3422c5c88ed026aa97cf4047db3d22;
modified for Tenacity)
Signed-off-by: Avery King <avery98@pm.me>