Remove time shift tool
The fact that Saucedacity has clip handles for moving individual audio
clips around in the track area has rendered this tool irrelevant. This
is rather identical to what Audacity has done upstream.

NOTE: as a behavioral difference from Audacity's changes, the Multi
Tool's shortcut key is still F6 rather than F5 as it is in Audacity.