Decrease device toolbar proportions
This kind of helps with an issue where the device toolbar drop-down
menus are too packed together. Unfortunately, this commit does NOT solve
that issue. There might have to be a re-work of this toolbar (e.g. a new
device toolbar) to fully solve this issue (well, depending on what
happens).

Pursuant to #23 in our discussions, some code has been marked as
obsolete but not yet dropped. This might be dropped in the future.