Neutralize some wxWidgets code in audio IO code
This is the start of the migration to Qt from wxWidgets. The first step
being taken here is to neutralize backend code, such as audio IO code.
This was initially started with AudioIO and is mostly independent of
wxWidgets, although not completely. RealtimeEffectManager, however, is
now entirely independent of wxWidgets.

There are clearly more spots to work on, but this is a (very) small
start. This is out of many bigger changes that will eventually come, as
will be standard when switching toolkits in a program like this.