Use lib-ffmpeg-support from Audacity
This version of lib-ffmpeg-support from Audacity is the version that was
first introduced. It does not necessarily come with FFmpeg support
(although, according to the commit when this was first added, "versions
55 and 58 are supported now").

Unlike in Audacity, lib-ffmpeg-support is not built as a library. It is
instead built into Saucedacity's core and the source is stored in
src/ffmpeg.

There have also been some miscellaneous changes too. Some subdirectories
in the Saucedacity source code: ffmpeg', 'shuttle', 'widgets', and 'xml'
have been added as include directories. This is namely to make including
some of these files easier all across Saucedacity's core.

This commit fixes #22, which involved avcodec_encode_audio2 returning
-22. Funny coincidence indeed.