Use Tenacity's libmad fork
I recently discovered that libmad 0.15.1b, the latest upstream version
of libmad, was released some time ago in 2004. From the time of this
commit (on July 24, 2022), that release was 18 years ago. Since then,
several CVEs have been listed for upstream libmad, likely to never be
fixed in the future.

Audacity's libmad fork had been used previously, but this was through
Conan. We plan to remove Conan entirely from Saucedacity, although it is
not guaranteed that these changes are going to be completed by 1.2
stable. There is no guarantee for when this will be finished, but things
are starting to come along as we progressively resolve our dependencies.

Tenacity's libmad fork introduces several changes in libmad, notably
fixes for existing libmad CVEs and a new CMake build system. This makes
Tenacity's libmad a more than perfect replacement for Audacity's libmad
fork (which resolved through Conan).

Despite these new developments, however, we will still allow for
building Saucedacity against the system's libmad. In some cases, like on
FreeBSD, using a local libmad is quite redudant because you could
Tenacity's fork already installed on your system. On the other hand, you
might prefer your system's libmad for other reasons. With all that said,
we will also still support libmad 0.15.1b as the minimum required
version for now.

Tenacity's libmad has been added as a git submodule, so in case there's
a new version of libmad, we can update it using git. This is in addition
to some other tricks we can perform too.