Fix libmad being configured improperly
It appears that in CI builds "mad.h" was not found when building
Saucedacity. This appeared to be due to the fact that libmad was not
being configured in libmad/CMakeLists.txt. (I was able to reproduce this
locally because the "Using FPM [...]" messages were absent from CMake's
output.

Hopefully this fixes CI builds. If not, at least libmad's FPM
configuration now works on our end.