Remove A LOT of references to Audacity
A lot of references have been removed to Audacity, but there are still
some that remain. They might be dealt with in some other form another
time.

Most of this was done using 'find' and 'sed'. There was no way I
could've done this all manuallly within a short amount of time.