Cleanup data dir code; remove more Audacity refs
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
!!                                                                           !!
!! There is one important change here: we no longer use ~/.Saucedacity-data  !!
!! on Linux. We now use XDG_CONFIG_HOME/saucedacity or                       !!
!! $HOME/.config/saucedacity instead.                                        !!
!!                                                                           !!
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!

A redundant #ifdef AUDACITY_NAME was removed. This is defined almost all
the time regardless of platform in CMake, so it should always exist
(unless we drop it for some reason). This code would use the name
'Saucedacity' instead of formatting them with AUDACITY_NAME. This code
has been removed.

AUDACITY_NAME was changed to APP_NAME, which is intended to be a more
neutral name, especially if anyone wants to fork Saucedacity to make
their own derivative.

More Audacity refs were removed from the source code in the process.
However, one change was reverted because it caused additional issues on
macOS (CMake config would fail due to an icon not being found due to its
name being accidentally changed).

Signed-off-by: Avery King <avery98@pm.me>