Drop support for older libav versions
==!! IMPORTANT CHANGES HAVE BEEN MADE IN THIS COMMIT !!==

The following versions have been dropped for their corresponding libav
libraries:

* libavcodec 55
* libavformat 55
* libavutil 52

This means that FFmpeg 2.x will no longer be supported. On Linux, this
means literally nothing. On Windows, however, this means that any
versions of FFmpeg for Audacity **will no longer be supported**. They
may still work, but there will be no support for using these older
FFmpeg libraries.

======================================================================

This is a part of an effort to bring better support for modern FFmpeg
versions in Saucedacity. Part of this includes dropping support for
older versions of FFmpeg, including FFmpeg 2.x and FFmpeg 3.x (and their
corresponding libav versions). FFmpeg 3.x will still be supported in
Saucedacity for the sole purpose of Ubuntu 18.04 until it's mainstream
support EOL date in April 2023. **There may be a decision to drop FFmpeg
3.x sooner than this date**.

Signed-off-by: Avery King <avery98@pm.me>