New library: lib-xml
Anything that was under src/xml is now in lib-xml. We no longer use
expat directly and instead use lib-xml in place of expat.

This library is mostly similar to Audacity's lib-xml, except for a
couple of files that are absent and the splitting of XMLFileWritter and
XMLStringWritter into their own source files.

Signed-off-by: Avery King <avery98@pm.me>