Bump Audacity wxWidgets version.
Audacity's latest version of wxWidgets available through Conan is
actually version 3.1.3.3-audacity. We are interested in using later
versions of upstream wxWidgets, however.

For the time being, if you are building Saucedacity with the default
settings, wxWidgets 3.1.3.2-audacity will now be used instead of
wxWigets 3.1.3.1-audacity. The main change that I could see is the
ability to build for ARM. This version of wxWidgets (using conan) is
also required if one is cross-compiling for the Raspberry Pi (I was
attempting this at one point; using wxWidgets 3.1.3.1-audacity did not
work while using 3.1.3.2-audacity did work).

If you were building against the system libraries before, you do not
have to worry about this change. In fact, from this point onward, we
recommend building against system libraries wherever possible on Linux.
On other platforms, sticking with conan is the best solution right now.

Signed-off-by: Avery King <avery98@pm.me>