Fix bug #101
When Tenacity crashes, semaphores are not necessarily cleaned up. At the
same time, new cliehnt processes assume that a server process is already
running and then try to connect to the non-existent server process.

To fix this, we introduce a new semaphore: /TenacityFailsafe. If we fail
to create this semaphore, we alert the user that Tenacity has previously
crashed and that the projects recovery dialog will open once Tenacity
has finished launching.

Reference-to: https://codeberg.org/tenacityteam/tenacity/issues/101
Signed-off-by: Avery King <avery98@pm.me>