Reapply commit d55cf9a34d
NOTE: I made some changes to comments here, so it's not a full 1:1
reapplication. However, this still fixes an issue with Tenacity failing
to open projects created in older versions of itself.

--------------------

Restore compatibility with Audacity-created projects

PR #84 introduced new project compatibility code taken from upstream.
However, this project compatibility code was geared towards focusing on
dealing with newer and older versions of Audacity, not Tenacity. This
resulted in Tenacity failing to open projects it created in prior
versions because of incorrectly assuming that the project was created in
a newer version of Tenacity when this was not the case. To fix this, the
following was introduced:

1. Add new version constants to represent Tenacity's version.

2. Check the project version against the Audacity version constants.
   This will succeed even if the project was created or updated in
   Tenacity.

While the solution is not ideal, it provides a temporary fix.
Unfortunately, we do not have a clear way of distinguising ourselves
from Audacity projects right now, but there might be ways that we can
distinguish ourselves from Audacity projects in the future.

Signed-off-by: Avery King <avery98@pm.me>
(cherry picked from commit d55cf9a34dca6c39ba845b9e6cf047917fe89c85)
Signed-off-by: Avery King <gperson@disroot.org>