Move export options into a seperate dialog
This was cherry-picked from commit a3dc2e0921. For whatever reason,
likely due to an error while rebasing, this commit never made it when
https://codeberg.org/tenacityteam/tenacity/pulls/110 got merged.

This commit is a little different than from the original. It shows the
export options dialog after alerting the user about an existing file (if
applicable) while the original commit did this before.

This is quite disappointing, but was luckily caught just before 1.3 beta
was released...

Signed-off-by: Avery King <avery98@pm.me>