Implement basic dynamic theming
This commit implements basic handling when the system's light/dark mode
preference is changed. The following was done:

* Add a new preference under 'Interface' named 'Respect system
  light/dark mode'.

* Add handling in ProjectWindow for handling a switch from light to dark
  mode when Tenacity is running.

There are notable limitations, however:

* Tenacity MUST be running in order for dynamic theming to work.

* Dynamic theming overwrites your current theme preference. If you are
  using the (classic) Audacity theme, Saucedacity theme, or high
  contrast theme, for example, your preference

Credit goes to @n0toose for coming up with the original POC for dyanmic
theming. While no code was taken from the original POC, they served as
an inspiration to make their POC (with enhancements) a reality.

Signed-off-by: Avery King <avery98@pm.me>