AboutDialog: Re-add include RevisionIdent.h
The include RevisionIdent.h statement was removed in commit 5ec0166529
but AboutDialog.cpp still references REV_LONG and REV_TIME defined in
RevisionIdent.h which isstill generated by the build system.

This commit restores the missing include so that revision id and timestamp
can be shown in the about box again.

Signed-off-by: David Karoly <david.karoly@outlook.com>