nyquist: fix Wide Path search in UNICODE Windows builds
xlisp gives pathes as char* that are using the Windows system codepage.
When building with UNICODE the default FindFirstFile uses FindFirstFileW, the
wide char version. This is likely going to crash or find nothing that should
be found.

We need to switch to the ANSI version of FindFirstFile which will use the
system codepage.

Signed-off-by: Steve Lhomme <slhomme@matroska.org>