Update translations for Tenacity 1.3
This is the first major update to Tenacity's translations. I'm hoping
that this change does not mess up our translations and the progress made
by existing translators. At the same time, I'm hoping that all source
strings are up-to-date and correct now and that Weblate still works, but
we'll see.

Signed-off-by: Avery King <avery98@pm.me>