Add an analyzer for clicks
It's similar to the Click Removal but instead of removing the clicks, it marks
them as label. This is a way to show the effect of the parameters
and if there are some missing or non clicks.

Later a companion plugin can remove only the clicks found in the labels,
after manual editing.

Signed-off-by: Steve Lhomme <slhomme@matroska.org>