Simply enables the cmake policy CMP0127
Alternative was to increase cmake required version to 3.22.  It could be successful, but that is a larger change that I wanted to avoid with no understanding.

Even with this method, the specific error message disappeared, and still found the FLAC* libraries.  So why did it trigger before for this specific statement?  Because of the plus signs?

Just making the minimal changes.