Move Karaoke into its own module
Audacity 3.5 removed Karaoke, presumed for its rather niche/demo role.
Therefore, move Karaoke into its own module so it can begin a new life
in Tenacity.

Karaoke won't see as many fixes as regular Tenacity, and it's now
ultimately possible to enable or disable it. At one point, this feature
will likely get dropped, but for now, it'll be kept as a module.

Signed-off-by: Avery King <avery98@pm.me>