Remove Audacity 2.x preference migrations
The pre-Audacity 2.4.0 preference migration adversely affected our
toolbars, namely the time toolbar by accidentally overwriting DockV2
with a value of '2', making Tenacity assume it was docked at the bottom
when it actually wasn't.

I have not observed any problems with the pre-2.2.0 config migration
code, but as a precaution, I also removed it because it was being
triggered anyways.

Signed-off-by: Avery King <avery98@pm.me>