Windows: Use dark mode title bar
Unfortunately, this didn't do much than I thought it would do, but since
this is just a proof-of-concept, I'm fine with these changes for all
intents and purposes.

This commit implements using a dark title bar in the project window.
Unfortunately, in wxWidgets 3.2.x, there aren't any dark mode APIs we
can use, so we have to resort to using the native APIs themselves on
Windows.

In wxWidgets 3.3+, which is still under active development at the time
of authoring this commit, there are new APIs related to using dark mode.
However, dark mode for Win32 apps is still experimental, and in
wxWidgets this uses an undocumented, and Microsoft-unsupported, API. One
of the new APIs surrounding application appearance in 3.3 appears to be
wxApp::SetAppearance().

Ultimately, I have to say that Win32 dark mode support and dynamic
theming are going to be limited with native controls anyways. However,
this will definitely be a plus for other platforms.

Signed-off-by: Avery King <avery98@pm.me>