ImportMka: Check for EbmlDummy object
Sometimes, it is possible for FindNextID() to return an EbmlDummy object
instead of a null pointer. In that case, check if the returned object is
a dummy object to avoid using it.

This also fixes a crash when attempting to import certain OGG files on
Windows (#311)

Reference-to: https://codeberg.org/tenacityteam/tenacity/issues/311
Co-authored-by: Steve Lhomme <slhomme@matroska.org>
Signed-off-by: Avery King <avery98@pm.me>
(cherry picked from commit 1a943aa97405b6372a0715ad08ef29841478c600)