ImportMP3: Skip unsigned long conversion if applicable
In our fork of libid3tag, and certain packages of 0.15.1b with
libid3tag-0.15.1b-64bit-long.patch applied, id3_ucs4_t can actually be 4
bytes where we expect it to be 8. If it's 4 bytes, we can skip an
unnecessary conversion and instead directly convert it to a wxString.

This comes from an ABI incompatibility on Arch Linux for #428. While
this commit does not fix the issue (the package needs to be recompiled),
I figured there could be an optimization by omiting a conversion since
we wouldn't need to use it against our fork.

Reference-to: https://codeberg.org/tenacityteam/tenacity/issues/428
Signed-off-by: Avery King <avery98@pm.me>
(cherry picked from commit 6d1677fc16ca112bc43ec5aa212714f47df27a21)