Mac: Remove wrapper
This wrapper was introduced with the intent of "[giving] the user more
control over where libraries such as FFmpeg get loaded from". It does
this by...unsetting DYLD_LIBRARY_PATH.

My view on this is to instead give this control to users *through*
DYLD_LIBRARY_PATH because it shouldn't be set under normal
circumstances. (It's another story if it is set, especially out of
necessity for something to work, for example).

Maybe we can add a note to the manual if it becomes a problem.

Signed-off-by: Avery King <avery98@pm.me>