vcpkg: Use a patched stable version of PortAudio
This commit updates vcpkg to using a stable version of PortAudio along
with using the stock PortAudio port. The only additional change to this
port was to add a patch enabling WASAPI loopback support. This is still
temporary until PortAudio 19.8 is released, but should be more
maintainable than having to rely on a git version of PortAudio that
could potentially break altogether at any time.

Signed-off-by: Avery King <avery98@pm.me>