Add notice about rebase
This is a temporary notice about our massive rebase effort. Pull
requests based off 'main' will no longer be accepted, only if they are
based against 'audacity-3.7-rebase' (the rebase branch). Thus, add a
note to the README and contributing policy about it.

Once the rebase is complete, and we merge 'audacity-3.7-rebase' into
main, we will continue working on the main branch as normal.

Signed-off-by: Avery King <avery98@pm.me>