TCP: Partially restore old button layout
The track title is now a button again that features the drop down.
However, the select button hasn't been restored, and the minimize button
remains in the top row of the track control panel. Additionally, the
track options menu does not contain an arrow pending a design decision.
(I'll see what happens when we release an alpha).

Signed-off-by: Avery King <avery98@pm.me>