OverlayPanel: Refactor DrawOverlays to use wxPaintDC
Providing a DC to DrawOverlays() is still optional, but it now
temporarily binds a custom paint handler to redraw the overlays.

Note: if you go to Help > Diagnostics > Frame Statistics, you'll notice
that the track panel is constantly updating now. This is because of the
track panel's timer, which continuously redraws the overlays. Later,
I'll see if there's a way I can refactor TrackPanel's drawing code to
not use a timer, but that might be a lot easier said than done.

Signed-off-by: Avery King <gperson@disroot.org>