Reapply commit e8b6f11573
As long as there are tracks present, draw edit cursor, scrub cursor and
quick play cursor across the entire track area.
Fixes problems introduced by Track Affodances, and also makes the
cursors behave like the play and record cursors.

--------------------

The edit cursor now draws across the entire track panel. However, there
appears to be a bug where if a track other than the first one is
selected, there is a bit of a gap between the track panel and the ruler.

Co-authored-by: akleja <storspov@gmail.com>
Signed-off-by: Avery King <gperson@disroot.org>