CMake: Add option to adjust directories on FreeBSD
Add -DFREEBSD_ADJUST_DIRS to allow the build to succeed on FreeBSD. Also
document it in BUILDING.md.

I don't suspect that any FreeBSD packages for FreeBSD don't need this
option, and in that case they can disable it if needed. I also don't
believe it interferes with any packaging attempts, but again, it can be
disabled easily if it does.

Signed-off-by: Avery King <gperson@disroot.org>