Add rebased dynamic compressor base class
This class is based on the old EffectCompressor2 class with all the
GUI-related bits stripped out. It is now based on StatefulEffect and has
been updated to conform to the new effect interfaces also.

Perhaps the most significant thing that's changed is the fact that realtime
parameters are updated in RealtimeInitialize() rather than UpdateUI(). We
no longer carry a sample rate member variable, so we can't just update it from
the UI anymore.

Signed-off-by: Avery King <gperson@disroot.org>