TrackPanel: Only use timer for active AudioIO
Commit cf4406d025 introduced more fixes for overlay drawing under
Wayland again, but it also caused the track panel to constantly update
as its timer was continuously redrawing the screen. However, it turns
out that most updates are already event-driven by themselves except for
audio playback. To help minimize this issue, we no longer use a timer
for general updates and instead use it when playback is active. This
way, Tenacity should run much better when idling and doing other tasks.

Signed-off-by: Avery King <gperson@disroot.org>