Introduce TrackControlPanel
This class is a base class that's responsible for holding all fo
Tenacity's track controls. The specifics will come later, but for now,
we'll focus on getting the basic controls down.

Currently, TrackControlPanel spans across the entire TrackPanel with a
red background. The first is unintentional as there will be another
class introduced to fill the portion that the red space will take.
Meanwhile, the second is intentional for now as it's for debugging
purposes.

No controls actually work yet. That will come later once we complete
this rather basic class.

Signed-off-by: Avery King <gperson@disroot.org>