TrackPanel: Don't check if the timer is active
This fixes a bug when clicking on the ruler to move the playhead at a
different position in playback. The check also wasn't necessary in the
first place because the timer would've been restarted anyways.

Also remove an incorrect comment, as AudioIOEvent::on tells you whether
or not the event that occurred (e.g., playback, capture, etc.) was
either turned on or off.

Signed-off-by: Avery King <gperson@disroot.org>