Remove reference to closed wxWidgets issue
I recently tried this and under wxWidgets 3.2.8.1 on Arch Linux under
KDE Wayland, I could not reproduce the issue. Neither could vadz either
under Sway, so the issue was closed.

This isn't a functional change, but I'm trying to remove references to
bugs that we likely won't encounter because of our ever increasing
wxWidgets minimum requirement. At least, if it can't be reproduced under
wxWidgets 3.2 or later, then it's gone.

Signed-off-by: Avery King <gperson@disroot.org>