Fix scrubbing
ScrubbingHandle now starts scrubbing when dragging and after a mouse
button release, as long as audio is not playing and the user didn't
previously drag the mouse. We also have to start playback in that
scenario as well to prime scrubbing.

There is still more work needed (we need to refactor the scrubbing UI
code to not rely on the track panel's timer), but for now, this appears
to restore scrubbing like before without going too crazy on updates.

Signed-off-by: Avery King <gperson@disroot.org>