TrackPanel: Initialize new members
Commit 32e5e34ad5c7229acba2c8a78cda528a6ad34232 introduced
new members to TrackPanel class, mIsPanning and mLastPanX,
but they were not initialized when TrackPanel instances were
constucted. Because of this, mIsPanning could be initialized
to something true-ish, which manifested itself as scrolling
to random place when trying to select part of the track in my
Audacity builds.

Fixes: 32e5e34ad5c7229acba2c8a78cda528a6ad34232
Signed-off-by: Ivan A. Melnikov <iv@altlinux.org>
(cherry picked from commit 661601e144449d1eeda13b6ad1edc485bb9daa28)